requirements.txt
streamlit>=1.36,<1.39
pymupdf>=1.26
pillow>=10.3